Move get_name_and_email() and test it.
